TL-10778 seminar: Fix attendees display if attendee exists in other event in add attendees
Reviewed-on: https://review.totaralms.com/13006
Tested-by: Jenkins Automation <jenkins@totaralms.com>

+++ mod/facetoface/attendees/addremove_html.php
@@ -21,8 +21,10 @@ $idx = 0; // Iterator to put elements on their positions when adding/removing.
         <?php
             if (!empty($attendees) && $action === 'add') {
                 foreach ($attendees as $attendee) {
-                    $fullname = fullname($attendee, true);
-                    echo "<option disabled value=\"$attendee->id\">".$fullname.", ".$attendee->email."</option>\n";
+                    if (!array_key_exists($attendee->id, $userstoadd)) {
+                        $fullname = fullname($attendee, true);
+                        echo "<option disabled value=\"$attendee->id\">".$fullname.", ".$attendee->email."</option>\n";
+                    }
                 }
             }
             if (!empty($userstoadd)) {
